+++ lib/plugin.dart
@@ -6,13 +6,11 @@ library stream_plugin;
 import "dart:async";
 import "dart:io";
 import "dart:collection";
-import "dart:mirrors";
 
 import "package:path/path.dart" as Path;
 import "package:charcode/ascii.dart";
 
 import "package:rikulo_commons/util.dart";
-import "package:rikulo_commons/mirrors.dart" show ClassUtil;
 import "package:rikulo_commons/io.dart" show getContentType;
 
 import "stream.dart";