+++ lib/src/server_impl.dart
@@ -128,7 +128,10 @@ class _StreamServer implements StreamServer {
 
     bool shouted = false;
     connect.errorDetail = new ErrorDetail(error, stackTrace);
-    var handler = _router.getErrorHandler(error);
+
+    var handler;
+    if (error is int) handler = _router.getErrorHandler(error);
+
     if (handler == null) {
       if (error is! HttpStatusException) {
         _logError(connect, error, stackTrace);
@@ -145,9 +148,8 @@ class _StreamServer implements StreamServer {
         _logError(connect, ex, st);
       }
 
-      handler = _router.getErrorHandlerByCode(code);
-      if (handler == null)
-        return;
+      handler = _router.getErrorHandler(code);
+      if (handler == null) return;
     }
 
     try {