+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # CHANGES
 
+### 2.5.0
+
+* [86](https://github.com/rikulo/stream/issues/86): the error mapping can map only error codes (integers). It implies
+    1. You can't specify the class name in the error mapping.
+    2. The type of the `errorMapping` arugment becomes `Map<int, dynamic>`
+    3. You'd like to handle a particular exception specially. You shall throw a special integer, say, -900, and then map it to your handler.
+
 ### 2.3.0
 
-* Fix [85](https://github.com/rikulo/stream/issues/85) The method '_ReadOnlyHeaders.add' has fewer named arguments than those of overridden method 'HttpHeadersWrapper.add [Dart 2.8]
+* Fix [85](https://github.com/rikulo/stream/issues/85) The method `_ReadOnlyHeaders.add` has fewer named arguments than those of overridden method 'HttpHeadersWrapper.add [Dart 2.8]
 
 ### 2.2.2+1